+++ codex/serializers/admin/__init__.py
@@ -1,0 +1,1 @@
+"""Admin view serialzers."""
