+++ codex/views/admin/library.py
@@ -18,7 +18,7 @@ from codex.librarian.watchdog.tasks import (
 )
 from codex.logger.logging import get_logger
 from codex.models import FailedImport, Folder, Library
-from codex.serializers.admin import (
+from codex.serializers.admin.libraries import (
     AdminFolderListSerializer,
     AdminFolderSerializer,
     FailedImportSerializer,
